Ajustes Ajustes no controller de Atividades para as regras de negócio da aplicação Ajustes no controller de Usuario para fazer login e retirar as seções Ajustes de Timezone para colocar o time zone correto no banco de dados e na aplicação Ajustes finais na tela de home para adicionar,editar e excluir Atividades e em login para suportar valores de reentrada de seção

+++ app/controller/UsuarioController.php
@@ -9,21 +9,32 @@ use App\Models\Usuario;
 use Illuminate\Database\Capsule\Manager as Capsule;
 use App\Core\RouteCore;
 use App\Models\Session;
+use DateTime;
 
 class UsuarioController extends Controller
 {
-    public function home()
+    public function home($error = null)
     {
         session_start();
         if (isset($_SESSION['usuario'])) {
-            Session::firstWhere('hash_session', $_SESSION['usuario']['session'])->delete();
+            $usuario = Session::firstWhere('hash_session', $_SESSION['usuario']['session']);
+            $diferenca = (new DateTime())->diff((new DateTime($usuario['start_at'])));
+            if ($usuario['remember'] != true && $diferenca->h >= 4) {
+                $usuario->delete();
+            } else if ($diferenca->d >= 1) {
+                $usuario->delete();
+            } else {
+                $this->redirect(Routecore::getRouteName('entrada'));
+                die();
+            }
+            $this->load('conta/login', ['usuario' => $_SESSION['usuario']['login']]);
+            session_destroy();
         }
-        session_destroy();
-        $this->load('conta/login');
+        $this->load('conta/login', [], $error);
     }
     public function login()
     {
-        $request = new Input;;
+        $request = new Input;
         $usuario = Usuario::where('usuario', '=', $request->post('usuario'))
             ->orWhere('email', '=', $request->post('usuario'))->first();
         if ($usuario != null && password_verify($request->post('senha'), $usuario['senha'])) {
@@ -31,9 +42,7 @@ class UsuarioController extends Controller
             session_start();
             $_SESSION['usuario'] = [
                 'session' => $hash,
-                'nome' => $usuario['nome'],
-                'usuario' => $usuario['usuario'],
-                'email' => $usuario['email']
+                'login' => $request->post('usuario'),
             ];
             Session::create([
                 'usuario_id' => $usuario['id'],
@@ -42,13 +51,15 @@ class UsuarioController extends Controller
             ]);
             $this->redirect(Routecore::getRouteName('entrada'));
         } else {
-            $this->load('conta/login', ['error' => 'Usuário ou Senha Invalidos']);
+            $this->home('Usuário ou Senha Invalidos');
         }
     }
     public function logout()
     {
         session_start();
-        Session::firstWhere('hash_session', $_SESSION['usuario']['session'])->delete();
+        if (isset($_SESSION['usuario'])) {
+            Session::firstWhere('hash_session', $_SESSION['usuario']['session'])->delete();
+        }
         session_destroy();
         $this->redirect(RouteCore::getRouteName('home'));
     }
